learning-recruter/fastapi-backend/requirements.txt
alembic
uvicorn
annotated-types
anyio
bcrypt
beautifulsoup4
cachetools
certifi
cffi
charset-normalizer
click
colorama
cryptography
dnspython
ecdsa
email_validator
fastapi
google
google-api-core
google-api-python-client
google-auth
google-auth-httplib2
google-auth-oauthlib
google-generativeai
googleapis-common-protos
greenlet
h11
httplib2
idna
Mako
MarkupSafe
oauthlib
passlib
proto-plus
protobuf
psycopg2-binary
pyasn1
pyasn1_modules
pycparser
pydantic
pydantic-settings
pydantic_core
pyparsing
python-dotenv
python-jose
requests
requests-oauthlib
rsa
six
sniffio
soupsieve
SQLAlchemy
starlette
typing-inspection
typing_extensions
uritemplate
urllib3
twilio
langchain
langchain-google-genai
PyPDF2
composio-core